+++ packages/coreflow_design_system/lib/coreflow_design_system.dart
@@ -42,6 +42,7 @@ export 'src/bold_scheme.dart';
 export 'src/bold_radius.dart';
 export 'src/bold_vidro.dart';
 export 'src/bold_tema_material.dart';
+export 'src/bold_pegador.dart';
 export 'src/bold_ponto.dart';
 export 'src/bold_pontos_de_pagina.dart';
 export 'src/bold_vinho.dart';